euler 20 and more intelligent testcases archiving

+++ hr_github.sh
@@ -20,6 +20,12 @@ COLOR_END="\033[0m"
 
 cmd_testcases_archive()
 {
+    # do not rebuild the archive if no testcases have been downloaded since last build
+    if python3 hr_count.py --latest; then
+        echo -e "${COLOR_LIGHT_PURPLE}Testcases up-to-date${COLOR_END}"
+        return
+    fi
+
     echo -e "${COLOR_LIGHT_PURPLE}Preparing testcases...${COLOR_END}"
 
     rm -rf tmp
